move init to use tel package
